test: 클래스 수정에 맞게 테스트코드 수정
